Invert `ObjectStore` retry logic (#858)
This PR is a follow-on to #819. That PR was very conservative about what
errors we retried on. Consequently, we continue to see all kinds of
errors that cause the DB to go into a dead state (e.g. `Os` connection
errors). The PR also used `std::io::Error`'s `TimeOut` instead of
`reqwest::Error`'s.

This PR takes a different approach: retry any error _except_ errors that
we know are permanent (even after manual intervention by an operator to,
for example, change ACLs on a bucket).

I did consider using `slatedb::ErrorKind`, but opted against it. We
currently treat `ObjectStoreError` as a `Error::operation`. This seems
right to me, but would result in us not retrying on a lot of transient
errors (ObjectStoreError seems to mixes transient and permanent errors).

I also opted to retry both table and manifest object store operations
for `usize::MAX`. This ensures that the process will recover when the
object store returns to a healthy state, no matter how long it takes. I
considered exposing this as a config but opted for `CLEAN_SLATE.md`
config style.

I should _also_ note that the `ExponentialBuilder` does allow for jitter
using an RNG. It takes a seed, which we don't supply. I've checked and
the builder has a [default
jitter](https://docs.rs/backon/latest/backon/struct.ExponentialBuilder.html#default)
set to `false`. This means we should be fine using this library with
DST, since the builder is deterministic by default.