Add `ClockedObjectStore` to mock times in DST (#862)
This PR adds `ClockedObjectStore`, which allows the DST to use its
`SystemClock` for `last_modified` metadata in the object store. This
change allows the garbage collector to see timestamps from the object
store that are relative to the mocked DST rather than the actual machine
clock. This change was largely vibe coded. I took a pass on it and made
a few tweaks; let me know if you spot anything odd.

While testing, I also noticed:

1. We have a divide-by-zero error in the compactor stat logging line.
2. We were not using the system clock when...
   1. making ManifestStore in `builder.rs`,
   2. creating `DbReaders`,
   3. and creating `Checkpoint`s.

I fixed these issues in this PR. All tests pass and the DST now hovers
at around 5-6GiB rather than the 50GiB+ it was previously accruing in a
few minutes of execution.

Fixes #861
Fixes #754