Implement MessageHandler for CompactorEventHandler (#888)
Update the compactor to use `MessageDispatcher`. I tried to do this in
the least invasive way possible. The implementation creates the
`MessageDispatcher` inside `Compactor::run_async_task`. I like this
pattern because it makes it pretty straight-forward to create a
Compactor and run it outside of the DB (in a separate process/on a
separate machine).

Part of #776