Implement MessageHandler for WalBufferManager (#896)
Replace `do_background` logic with `MessageDispatcher`. I tried to keep
the changes as small as possible.

This PR was more involved than the previous migrations. WalBufferManager
is quite different and attempted to handle some of the logic that
MessageDispatcher now does. I'm leaving comments in this PR in areas
that need particular attention.

Review after #891
Part of #776