Expose fencing errors & mark `ErrorKind` as `non_exhaustive` and `Copy` (#901)
This does exactly what the title says.

Other options for fencing errors would be to add a `fenced` boolean
field to `ErrorKind::Permission`, or to rename that variant or guarantee
that it only applies to fencing. In any case users should be able to
tell whether an error is due to fencing, and in my opinion it shouldn't
be mixed with any other error kinds.

As for the other part of the changes, they were already discussed in
#892.

Fixes #881