Replace `BackgroundTaskShutdown` with a clean shutdown (#918)
A clean shutdown in SlateDB (`db.close()`) currently sets a
`SlateDBError::BackgroundTaskShutdown` in `db_state.rs`'s `error` field.
We then special case this error type all over the place to determine
when we're shutdown cleanly. This is confusing and brittle.

I've removed `BackgroundTaskShutdown`. `dispatcher.rs` now returns
`Ok(())` when it's shutdown cleanly. If a dispatcher run_loop returns
with an `Err(...)`, it still sets the `error` field in `db_state.rs`.
Doing so immediately halts all other dispatchers and blocks future calls
to the DB.

I originally thought we should convert `db_state.rs`'s `error:
WatchableOnceCell<SlateDBError>,` to be `shutdown_result:
WatchableOnceCell<Result<(), SlateDBError>>`. The idea was that we'd set
the field to `Ok(())` on shutdown to signal that the DB should be
shutdown. But I found this overlapped with `cancellation_token`'s
purpose. I kept `cancellation_token` since it seems idiomatic. I checked
if there's a pattern to include a `Result` with a cancellation token in
Rust, but there appears not to be. I landed on:

- cancellation_token() is used to request a clean shutdown
- dispatchers never halt unless
- an error occurred in the dispatcher (it will then halt and set the
`error` field)
  - the error field is set by another task (another dispatcher)
- the cancellation_token is called (Ok(()) is returned and no error is
set)

I did find that there were a few places where we were notifying channels
that we could no longer handle their request in `cleanup()`. I
introduced `SlateDBError::Closed` to signal to those channels that the
DB is now closed and their requests couldn't be completed. I did a quick
pass and it appears `Closed` is being used exclusively for `send_safely`
channel calls, which makes sense.

Fixes #817