Wait for both writes in test_should_tombstones_in_l0 (#923)
`test_should_tombstones_in_l0` failed in a recent nightly:

https://github.com/slatedb/slatedb/actions/runs/18001434698/job/51211440101

There is a race condition in this test where the compactor can see only
the first write when the flush is triggered. This is because the
compactor maintains its own view of the DB state (manifest)
independently of the writer. It refreshes this view periodically. It is
possible to have this sequence of events:

1. db.put(k1, v1) (line 730 in `compactor.rs`)
2. k1 put is flushed to l0 SST
3. writer updates manifest with l0 SST
4. compactor refreshes (`CompactorEventHandler::handle_ticker`) its
manifest and sees a single L0 SST
5. db.put(k2, v2) (line 731 in `compactor.rs`)
6. k2 put is flushed to l0 SST
7. db.flush() (line 732 in `compactor.rs` noop since the previous put's
were await_durable=true, thus guaranteed to already be flushed)
8. `scheduler.should_compact.store(true, Ordering::SeqCst);` (line
deleted now)
9. compactor runs, compacts the single L0 SST into SR=0, and flips
`should_compact` back to false
10. `await_compaction` is called (line 733)

At this point, before my changes, `await_compaction` would time out
because the compactor scheduler never sets `should_compact` to true
again. The k2 L0 SST will sit in L0, which violates the `empty_l0`
clause in `await_compaction`.

The fix for this is rather ugly. There's no good way to get at the
compactor in `db` since it doesn't hold on to it. Without that, there's
no way to force the compactor to refresh its DB state so that it sees
both L0 SSTs _before_ we set `should_compact` to true.

I ended up tweaking the OnDemand scheduler to take a function instead. I
then encoded the rules for the test in the function. This is pretty
gross. I briefly considered disabling the compactor in the `db` all
together, and instead creating one entirely separately in the
`test_should_tombstones_in_l0`, which would be cleaner. I opted not to
do this just to minimize the amount of changes since these tests are so
delicate as it is.