Don't double-cache object store in test helper (#921)
test_get_with_object_store_cache_stored_files was failing randomly on
the Windows cross build. I'm unable to reproduce the issue locally, but
I did notice that the helper the tests uses wraps the object store in
two `FsCacheStorage` and points them to the same directory.

I inspected the temp directory and found that it held both a 1kib and
4mib part file (one for each cache). I suspect there's a race between
these two caches that is causing the failure. I've updated the helper to
disable the internal object store cache and _only_ use the one we
explicitly set in the test.