move bloom filter access into the SST iterator (#942)
This PR moves the bloom filter access into the iterator so any SST
iterator created with just a single key in the range will be able to
leverage it. This is the next step in removing special casing of `get`
vs `scan` and unifying the code to use the same iterators.