fix merge iterator iteration order implementation (#947)
There's an assumption that iterators return values in increasing key
order and decreasing sequence number order. The implementation of the
merge iterator doesn't do this cleanly - the heap actually organized
things such that the first iterator returned has the lowest key AND
lowest sequence number.

This didn't surface as a bug during the normal operation because it was
masked when de-duplicating. It _is_ a bug, though, for compaction (the
test actually confirmed the wrong behavior, see `test_dedup_disabled`).

This change fixes that and also prepares this for the work supporting
merges, in that it will eagerly return values that are merges until it
finds the first non-merge. This starts to deviate from the approach in
the `MergeOperatorIterator` because merges need to be applied a bit more
carefully than the approach there suggests depending on whether values
are merged using MergeIterator or the GetIterator (which early-exits on
the first found key for Get operations).