support merges on the read path (#948)
this PR supports the read side of merges

Main changes in this PR:

- The "main" change is in `db_iter.rs`. Instead of directly using the
`ScanIterator` or the `GetIterator` we now optionally wrap it in a
`MergeOperatorIterator` which will resolve merge values if returned by
the underlying iterator
- The `MergeOperatorIterator` itself was reworked to fit in this context
and I fixed a bug where newer merge entries were merged first. Now it
collects all valid merges into a list from newest to oldest, then
reverses it before applying the merge operator to all of them.
- A boat load of test were added to `reader.rs`

Remaining TODOs:

- [ ] implement public APIs for inserting merge values
- [ ] implement public APIs for configuring the merge operator
- [ ] make sure the compaction side of things works properly (unify the
iterators with compaction as well)