support merges in WriteBatch (#950)
most of this PR is test code, the actual changes are:

1. some setup for wiring in the merge operator (unused in this PR, but
the changeset was getting large with just the changes to `WriteBatch` so
I wanted to submit this PR as is for now)
2. Change `WriteBatch` from `BTree<Bytes, WriteOp>` to
`BTreeMap<KVTableInternalKey, WriteOp>` so that it can accept multiple
writes for the same key
3. Simplify the iterator by wrapping it in `Peekable` so we don't have
to maintain `current` on our own

**NOTE:** I chose not to resolve merges on insert for `WriteBatch`
because that would require passing in the `MergeOperator` into
`WriteBatch` which could only really be done by changing the public API
for `WriteBatch`. Not only that, but it doesn't really simplify things
because we'd still need to store multiple operations for a single key
because you could `delete` a key and then `merge` and we'd still need to
store that both operations happened so we could flush a tombstone and
then a merge.