Remove `Box<dyn Any + Send>` payload from `BackgroundTaskPanic` (#965)
This is a follow-on to #944. In that PR, I added a
`MessageHandlerExecutor` that manages the lifecycle for all dispatchers.
I was planning to expose background task results/panics in a subsequent
PR. However, we decided against this because:

1. There's not much use to the information. If a BG task panics, the DB
goes into an error state and must always be restarted.
2. The changes to do so add some complexity.
3. Exposing internal panic payloads to the user can leak implementation
details.

Instead, I've opted to simply remove the panic payload from
`SlateDBError::BackgroundTaskPanic`. Now, two things happen when a BG
task panics:

- The panic payload is logged at `error!` level
- A `SlateDBError::BackgroundTaskPanic(task_name)` is returned

The latter will cause the DB to go into a `Closed(Panic)` error state.