Fix `test_compactor_compacts_l0` (#971)
I noticed a sporadic failure in `test_compactor_compacts_l0` in
https://github.com/slatedb/slatedb/pull/968, which was simply an RFC
change.

The `await_compaction` is using a split view of the DB state between
`db.inner.state` and `get_db_state`, which returns the persisted
manifest's view of the state. These two are not necessarily in sync, so
we can get into a state where the manifest l0 might be empty but the
in-memory l0 might not. This can lead `await_compaction` to think that
all data has made it to L1+ when, in fact, it has not. I fixed this by
making `await_compaction` use the in-memory view for everything.