Simplify compactor in-memory model (#981)
Let me start by saying that my primary goal at this moment is getting
0.9.0 out the door in a stable fashion. With that in mind, I've
dramatically trimmed down the compactor code to represent _only the
current functionality_. Much of the changes introduced in #954 and
before (as part of RFC13) have been deleted. This includes:

- Removing compactor.fbs entirely. It was completely unused.
- Removing `dead_code`'s
- Removing CompactorJobResponse, CompactorJobRequestType,
CompactorJobProgress, CompactorStateRecord, and CompactorJobInput,
CompactorJobStatus
- Renaming CompactorJobRequest -> CompactorJobSpec (requested by
@rodesai)
- Eliminating much of the compactor/compaction confusion by simply
eliminating such prefixes.
- Clarifying job_id vs. attempt_id by being explicit about param and
variable names _at all times_ except in the data structure itself
(CompactorJob::id is OK).
- Replacing `compactor_jobs` and `scheduled_jobs` with a single `jobs`
- Eliminate the 1 : N mapping between CompactorJob and
CompactorJobAttempt for now.
- Drop CompactorJobAttempt entirely from CompactorState. It remains an
implementation detail of the compactor, as it was.

These last few points are the most important. They're done in a372294.

I acknowledge, that we will probably end up re-introducing much of this.
But I hope it's in a more piecemeal fashion.