Prevent GC from deleting active compaction SSTs (#984)
Introduce a process-local coordination mechanism between the compactor
and garbage collector to prevent premature deletion of SSTs that are
still being compacted. A new "compaction low watermark" metric is
introduced, representing the earliest start time of all active
compaction jobs. The GC may not delete any compacted SSTs greater than
this timestamp (since the SST could be part of an active compaction job
that hasn't yet completed).

This is a hacky interim solution for #604 until persistent compactor
state is implemented. It won't work if the GC and compactor don't share
the same `StatRegistry` or if they're not in the same process/machine.