clean up TransactionalObjectStore
We introduced TransactionalObjectStore early on to allow us to
abstract the algorithm for doing atomic CAS operations against
object stores. This was important because S3 didn't support CAS
and so would require a custom implementation.

This is no longer required since S3 supports CAS, and the
object_store crate uses it to implement atomic create-if-not-exist.
So in this patch we drop the TransactionalObjectStore trait. The
impl (DelegatingTransactionalObjectStore) was doing 2 things: (1)
implementing txns by delegating to object_store and (2) providing
an object store like API but scoped to a path prefix. We still want
(2) but just for the manifest stuff, so that's moved to a wrapper
type called PathScopedObjectStoreAccessor internal to the record
module