rework the id stuff so next is done from store
This patch slightly tweaks how the version is set when writing an
object. Previously, we'd pass the next version to the ops. In this
patch we pass the expected current version and let the ops impl
select the next version. This allows us to simplify the trait bound
passed to the transactional objects to just TransactionalObjectOps