Add flatbuffer evolution check (#1008)
There are some upcoming changes that require modifying the flatbuffers
(compactor state, #779, #985, among others). I figure now's as good a time as any to
start to do some compatibility checks for our storage system. I'm starting very
gently in this PR with a simple [flatbuffer
--conform](https://flatbuffers.dev/evolution/#checking-conformity) check. This command
checks the following that fields are never removed, or
re-ordered, and that types evolve in a compatible way, and so on.