transactional object api (#1009)
## Summary

This patch renames the `RecordStore` stuff to `TransactionalObject`, but
more importantly it reworks it so
that it provides a standalone API for transactionally writing objects
using various kinds of backends/protocols. In the short
term this hopefully gives us a cleaner abstraction we can reuse for
things other than the manifest. Longer term,
this is a step towards being able to move some of these foundational
protocols that slatedb is built on out of the
main slatedb crate so they can be used independently.

This draft PR is more of a proposal for an API. If it makes sense to
folks I'll break it down into a few sub-PRs. These
map 1:1 with the commits in this patch
[ ] A simple rename the Record stuff to Transactional Object
[ ] Introduction of the various traits that define the interface to
transactional objects
[ ] A tweak to the `TransactionalObjectOps` interface to pass in the
expected id rather than the new id
[ ] A rename of the module to transactional_object
[ ] Splitting off the error type used by the transactional_object module
into its own error type

## Notes for Reviewers

For now I'd suggest just looking through transactional_object/mod.rs and
seeing if the API makes sense.

Thank you for the review! 🙏