set last_remote_persisted_seq_num after flushing memtable
before this patch the db writer wasn't notifying the oracle about
the last seq num in an l0 when the l0 is successfully written and
added to the manifest. This is a bug because at this point all
contained rows are durably stored to the db. It's also possible for
this to occur before those rows are durably written to the WAL,
so its important to notify the oracle when this happens so readers
using durability level Remote can observe the contained rows.