Support specifying a name when creating checkpoints
With a name associated with a checkpoint, one can list the checkpoints
by name. Note that the unqiueness of names across all the checkpoints
is not enforced. So, when a list operation is performed using a name
as a filter, then multiple checkpoints could get listed.

The following changes have been made:

[1] A new optional field called 'name' has been added to the flatbuffer
table called 'Checkpoint' to persist the name of the checkpoint. Respective
encoder/decoder have been updated to handle the name field.

[2] The admin API 'create_checkpoint' has been updated to accept 'name'
as an option. Corresponding backend changes have been done to ensure
that 'name' gets encoded into the flatbuffer.

[3] The admin API 'list_checkpoint' has been updated to accept a parameter
called 'name_filter' which will be used to filter only the checkpoints
that have that name.

[4] slatedb-cli and slatedb-py modules have been updated to pass along the
name field in both 'create_checkpoint*' and list_checkpoints APIs