filter out old sequenced keys when flushing to L0
When flushing immutable memtables to L0, this commit
filters out old sequenced keys. For each key, it only
flushes the entry with the highest sequence number in the
immutable memtable.

Merge entries are not filtered. Handling merge entries is
left to a future commit.

This commit resolves https://github.com/slatedb/slatedb/issues/1051.