StoredManifest init and create_new_db to accept system clock as param
Currently, the clock stored in ManifestStore is passed in to these functions
which in essence is an encapsulation violation. We would be updating both the
StoredManifest::init and StoredManifest::create_new_db functions to explicitly
accept a system clock parameter.

This change also improves testability by allowing tests to inject mock clocks
when needed.